requirements.txt
bili_ticket_gt_python~=0.2.7
requests~=2.32.3
Brotli~=1.1.0
httpx~=0.27.2
msgpack~=1.1.0
openpyxl~=3.1.2
PuLP~=2.9.0
pycryptodome~=3.19.1
pydantic~=2.9.2
PySide6~=6.7.2
ruamel.base~=1.0.0
ruamel.yaml~=0.18.10
SQLAlchemy~=2.0.35
tinydb~=4.8.0
UnityPy~=1.10.18
zstandard~=0.21.0
